add target tracker mode based on replay ghost
if replay doesnt exist for the pb, fall back to old behavior. if pb doesnt exist, fall back to goal percent (old behavior). you can use this to compare your pb to online scores 1:1 when watching replays. you can also use this to compare your pb to your own replays. this only works for your pb of the current chart for the current rate. no sense in pulling the pb of the wrong rate. changing judge also converts correctly, so you can set a j4 99.99 pb and replay it in j7 to go for higher quad/quint and it compares correctly